refactor: modularize codebase structure
- Split app.rs into app/ module (message.rs, osd_info.rs)
- Split tempo.rs into tempo/ module (calendar.rs, mod.rs, weather.rs)
- Extract DBus utilities into dbus_utils.rs
- Refactor theme module for cleaner structure
- Update services to use extracted utilities
