+++ src/config.rs
@@ -1,10 +1,10 @@
 use crate::app::Message;
 use crate::i18n::{UnitSystem, unit_system};
 use crate::services::upower::PeripheralDeviceKind;
-use crate::utils::celsius_to_fahrenheit;
+use crate::utils::{celsius_to_fahrenheit, send_or_log};
 use hex_color::HexColor;
 use iced::futures::StreamExt;
-use iced::{Color, Subscription, futures::SinkExt, stream::channel, theme::palette};
+use iced::{Color, Subscription, stream::channel, theme::palette};
 use inotify::EventMask;
 use inotify::Inotify;
 use inotify::WatchMask;
@@ -1336,9 +1336,11 @@ pub fn subscription(path: &Path) -> Subscription<Message> {
                                     .await
                                     .unwrap_or_default();
 
-                                    let _ = output
-                                        .send(Message::ConfigChanged(Box::new(new_config)))
-                                        .await;
+                                    send_or_log(
+                                        &mut output,
+                                        Message::ConfigChanged(Box::new(new_config)),
+                                    )
+                                    .await;
                                 }
                                 Some(Event::Removed) => {
                                     // wait and double check if the file is really gone
@@ -1346,9 +1348,11 @@ pub fn subscription(path: &Path) -> Subscription<Message> {
 
                                     if !path.exists() {
                                         info!("Config file removed");
-                                        let _ = output
-                                            .send(Message::ConfigChanged(Box::default()))
-                                            .await;
+                                        send_or_log(
+                                            &mut output,
+                                            Message::ConfigChanged(Box::default()),
+                                        )
+                                        .await;
                                     }
                                 }
                                 None => {
@@ -1374,4 +1378,54 @@ pub fn subscription(path: &Path) -> Subscription<Message> {
             }
         })
     })
+}
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_default_config() {
+        let config = Config::default();
+        assert_eq!(config.log_level, "warn");
+        assert_eq!(config.position, Position::Top);
+        assert_eq!(config.layer, Layer::Bottom);
+        assert!(!config.enable_esc_key);
+    }
+
+    #[test]
+    fn test_config_validation() {
+        let mut config = Config::default();
+        config.updates = Some(UpdatesModuleConfig {
+            check_cmd: "test".to_string(),
+            update_cmd: "test".to_string(),
+            interval: 0,
+        });
+        config.validate();
+        assert_eq!(config.updates.as_ref().unwrap().interval, 1);
+    }
+
+    #[test]
+    fn test_threshold_validation() {
+        let mut cpu = SystemInfoCpu::default();
+        cpu.warn_threshold = 90;
+        cpu.alert_threshold = 80;
+        cpu.validate();
+        assert_eq!(cpu.warn_threshold, 80);
+    }
+
+    #[test]
+    fn test_parse_module_name() {
+        let module: ModuleName = "Workspaces".parse().unwrap();
+        assert_eq!(module, ModuleName::Workspaces);
+        
+        let custom: ModuleName = "MyCustomModule".parse().unwrap();
+        assert_eq!(custom, ModuleName::Custom("MyCustomModule".to_string()));
+    }
+
+    #[test]
+    fn test_appearance_style_default() {
+        assert_eq!(AppearanceStyle::default(), AppearanceStyle::Islands);
+    }
 }
